jitterlisp bugfix: make remainder-by-two of a negative return a negative
My previous attempt and effectively returned the absolute value of the
remainder, ignoring the sign of the argument.

With a negative dividend I want to return a negative remainder, as C99 does.
This new solution, by myself, is not the most beautiful but at least is
branchless, and now correct.

* example-vms/jitterlisp/jitterlisp-operations.h (JITTERLISP_EXP_F_F_2REMAINDER):
Redefine macro, now conditionally according to JITTERLISP_FIXNUM_TAG's value.