make JITTER_EXIT a branching macro
From now on using JITTER_EXIT in a VM instruction requires the instruction to
be branching; this is necessary since exiting VM code entails a hardware branch
(to the end of the execution C function), which may be defective and therefore
needs to be compiled as a replacement as well.

Fix an example using JITTER_EXIT.

Discovered thanks to José Marchesi, who is using JITTER_EXIT in poke.

* jitter/jitter-executor.h (JITTER_EXIT) [replicate]: Rename to _JITTER_EXIT.
(JITTER_EXIT) [! replicate]: Rename to _JITTER_EXIT.
Add comment.

* jitterc/jitterc-generate.c (jitterc_emit_executor_definitions): Handle
JITTER_EXIT along with the other branching macros.
(jitterc_emit_executor_main_function): Use the private macro _JITTER_EXIT from
the special specialised instruction, which is not relocatable and therefore
safe to compile only once.
Fix generated comment.

* example-vms/uninspired/uninspired.jitter (endvm): Add branching attribute.