support -m32 on x86_64
On x86_64 check for pointer size at configuration: configure for i386 instead of
x86_64 if the size is 4B instead of 8B.
(This change set is different from the one with the same summary on the master
branch, even if it implements the same functionality)

Thanks to José Marchesi for reporting the issue.

* configure.ac (JITTER_ASSEMBLY_SUBDIRECTORY, jitter_supports_minimal_threading)
(jitter_supports_no_threading) [x86_64]: Add conditional.