handle long path names in tarballs from git branches
When generating a tarball from a non-master branch it is better to take the
liberty of relying on a tar program without pathname length limitation -- which
is checked for at configuration time.  This is a good idea now that the branch
name is part of the version identifier.

* bootstrap.conf: Copy machine/dummy--minimal-threading to machine/dummy.  Add
comment.

* configure.ac: Change comment.
(jitter_filename_length_limit_options): New M4 macro.
(AM_INIT_AUTOMAKE): Call with new macro.
(JITTER_ASSEMBLY_SUBDIRECTORY): Add comment about copied "dummy" directory.