structured: compile return statements as tail, in any position
A return statement should be considered tail even when not in a tail context.
Moreover its result expression is also in a tail context -- which may be
important in some cases, such as when the result is a conditional of which at
least one branch is a procedure call.

Perform the same change in both code generators, register and stack.

* example-vms/structured/structured-code-generator-register.c
(structured_translate_statement): Change definition and comments.
* example-vms/structured/structured-code-generator-stack.c
(structured_translate_statement): Likewise.