build system: no longer use info-in-builddir
No distcheck regressions.

* Makefile.am (AUTOMAKE_OPTIONS): Remove info-in-builddir .
(doc_jitter_TEXINFOS): Refer to doc/gpl.texi doc/fdl.texi without the $(srcdir)
directory, which is obvious; notice that this is not always the case for the
copies of these files included by other Texinfo documents, since the copies
pathname is explicitly used in make rules.