add minimal-threading support for s390x
The test suite passes with no errors and all the disassembled code I checked
looks reasonable.  This is a minimal-effort port for minimal-threading, which is
how things are supposed to be.
The older s390 variant is still not tested because I have no usable emulator for
it.

* machine/s390x/jitter/machine/jitter-machine.h: New file.

* machine/s390x/jitter/machine/jitter-machine-assembly.S: New stub file, copied
from the dummy file.
* machine/s390x/jitter/machine/jitter-machine-c.c: New stub file, copied
from the dummy file.

* Makefile.am (EXTRA_DIST): Distribute new files.
* configure.ac ($default_host_cpu case): Add s390x case.  Keep existing s390.