replication: before shrinking executable buffer check that its size went down
Add a defensive check to prevent disasters.  This should not be triggered in
production but it can happen on complex-dispatch (minimal-threading or
no-threading) VMs compiled with unsupported CFLAGS -- a combination which will
almost certainly not work anyway, but it is better to fail cleanly.

* jitter/jitter-replicate.c (jitter_replicate_program): Add a check.

This was reported by José Marchesi, working on GNU Poke.  Thanks.