pickles: add new pickle for EH Frame section
This commit adds a pickle for EH Frame section in ELF binaries.

At the moment, the blob of bytes containing the CFI insns for the FDE
(Function Descriptor Entry) die cannot be interpreted.  The CIE (Common
Information Entry) die, however, is fully represented.

This issue can be fixed in the future.

2021-06-03  Indu Bhagat  <indu.bhagat@oracle.com>

ChangeLog:

	* pickles/Makefile.am: Include eh-frame.pk.
	* pickles/eh-frame.pk: New pickle.