pickles: add pickles for CTF Frame section
This patch adds new pickles to read and manipulate the CTF Frame
section.  CTF Frame format is a lightweight and compact unwind format
designed to serve the usecases of fast, online stack unwinding. CTF Frame
format is currently being designed and discussed in the open source
GNU/Linux community.

This pickle will be adapted in the future as the format evolves.

2022-07-04  Indu Bhagat  <indu.bhagat@oracle.com>

	* pickles/Makefile.am: Include CTF Frame pickles.
	* pickles/ctf-frame-dump.pk: New file.
	* pickles/ctf-frame.pk: Likewise.