poke: Prefer ${XDG_CONFIG_HOME:-$HOME/.config} to XDG_CONFIG_DIRS
Some DEs (namely, KDE) set the latter without setting $XDG_CONFIG_HOME,
because the XDG specification says:

    Specifications may reference this specification by specifying the
    location of a configuration file as
    $XDG_CONFIG_DIRS/subdir/filename. This implies that:

    ...
    - A user-specific version of the configuration file may be created
      in $XDG_CONFIG_HOME/subdir/filename, taking into account the
      default value for $XDG_CONFIG_HOME if $XDG_CONFIG_HOME is not set.
    ...

It is implied (and also followed by other programs) that the default
value of XDG_CONFIG_HOME should be preferred to $XDG_CONFIG_DIRS, even
when the former is not set but the latter is. This also seems to be in
the spirit of the standard, which treats XDG_CONFIG_HOME as a writable
location (so, the program can modify it in accordance with the users
wishes), while treating XDG_CONFIG_DIRS as read only locations, normally
for default values or distro-provided overrides.

2022-09-24  Arsen Arsenović   <arsen@aarsen.me>

	* poke/poke.c: Prefer $XDG_CONFIG_HOME to $XDG_CONFIG_DIRS, even
	when unset.