libpoke: make assertions with no explicit msg show the source code of its condition
Most of the times the textual form of the condition in an assertion is
the best hint on seeing what is going on.  Up to now, we would need to
replicate the condition text like this:

  assert (1 != 1, "1 != 1");

This patch makes the second argument to `assert' to default to the
source code of the condition specified as the first argument.
Therefore, in most cases we just have to write:

  assert (1 != 1);

And we will get an useful hint:

  unhandled assertion failure exception
  <stdin>:1:1: 1 != 1

2022-10-03  Jose E. Marchesi  <jemarch@gnu.org>

	* libpoke/pkl-diag.h: Prototype for pkl_loc_to_source.
	* libpoke/pkl-diag.c (pkl_loc_to_source): New function.
	* libpoke/pkl-tab.y (pkl_make_assertion): Call pkl_loc_to_source.
	* pickles/pe.pk: Use asserts without explicit message.
	* pickles/color.pk: Likewise.
	* pickles/asn1-ber.pk: Likewise.
	* pickles/mcr.pk: Likewise.
	* pickles/rgb24.pk: Likewise.
	* testsuite/poke.pkl/assert-6.pk: New test.
	* testsuite/poke.pkl/assert-7.pk: Likewise.
	* testsuite/poke.pktest/pktest-11.pk: Adapt test.
	* testsuite/poke.pktest/pktest-7.pk: Likewise.
	* testsuite/poke.pktest/pktest-5.pk: Likewise.
	* testsuite/Makefile.am (EXTRA_DIST): Add new tests.
	* doc/poke.texi (assert): Document new behavior.