std: make stoca more convenient
The old signature of stoca made it cumbersome to use in simple
scripts. One would like to call it as

    stoca("Poke")

and obtain the array `['P', 'o', 'k', 'e']` in this way. The signature
of poke is modified to allow this; we maintain backwards-compatibility
with the old uses of stoca save for one case, where the user
explicitly relied on stoca throwing `E_out_of_bounds` for an empty
second argument.

We also modify the documentation to match the new signature, and add
the invariance test case which checks that stoca and catos are
inverses.