pickles: sframe.pk: update pickle to reflect AArch64 PAuth info
On AArch64, PAuth extension allows pointer signing and authentication.
SFrame format encodes whether the return address is signed; Also it
encodes which key is used for signing (Key A, or Key B) per SFrame FDE
on AArch64.

SFrame support in Binutils (2.40) includes the generation of this
information.  This patch updates the pickle to reflect the bits used for
conveying this information in the SFrame pickle.

Fixes some typos as well.

2023-02-07  Indu Bhagat  <indu.bhagat@oracle.com>

	* pickles/sframe.pk: Update the pickle for AArch64 PAuth
	support.