ios.pk: make dump :val hyperlinks with non-ephemeral references
The `dump :val' shorcut numbers (1, 2, ...) for value constituents are
ephemeral: their validity only spans until the next `dump :val'
command.  However, this is inconvenient when they are used as
hyperlinks.

This patch changes ios.pk to generate "closure" hyperlinks instead of
"execute" hyperlinks.  The closures keep a reference to the values, so
they are not ephemeral.

2023-08-24  Jose E. Marchesi  <jemarch@gnu.org>

	* pickles/ios.pk: Use "closure" hyperlinks to guarantee that they
	will work ever if another subsequent dump is issued.