pk-help.pk: hint at activating the pager in .help
It is to be expected that a new user will follow the recommendation
expressed in the program banner, to run .help.  The output of that
command is long and will likely scroll down.  This patch adds a
message to the output of .help recommending the user to activate the
pager with `.set pager yes'.

2023-10-30  Jose E. Marchesi  <jemarch@gnu.org>

	* poke/pk-help.pk (pk_help): Hint at activating the page at the
	end of the output of `.help'.