pkl: introduce a lexical pass
The Poke compiler builds the compile-time lexical environment
corresponding to the structure of the compiled program at parse time.
This is necessary in order to link identifiers with their
declarations, and also to distinguish between variables and type
names.  It also annotates variables with their lexical address, in the
form of a pair of `back' and `over' values.  Once the parser finishes,
this compile-time lexical environment is discarded with the only
exception of the top-level.

Then the resulting AST is subject to many analysis and transformation
phases.  Since the lexical addresses of variables have already been
assigned, these transformations shall be careful to either not
introduce structural changes in the program that would result in a
change in these addresses, or to update the addresses in an ad-hoc
manner.

This is no good.  This patch changes the parser in order to not
annotate lexical addresses (which remain "unknown") and adds a lexical
pass to the compiler front-end which runs right before code
generation.  This pass reconstructs a (possibly updated) lexical
environment and assigns lexical addresses.  This makes it possible for
other front-end passes to update the AST at will without having to
worry about the lexical structure.

2024-06-02  Jose E. Marchesi  <jemarch@gnu.org>

	* libpoke/pkl-trans.h (struct pkl_trans_payload): Add field `env'.
	(pkl_phase_transl): New extern.
	(pkl_trans_init_payload): Initialize env.
	* libpoke/pkl-trans.c (transl): New phase.
	(PKL_TRANS_ENV): Define.
	(pkl_transl_ps_var): New handler.
	(pkl_transl_pr_comp_stmt): Likewise.
	(pkl_transl_ps_comp_stmt): Likewise.
	(pkl_transl_pr_try_stmt): Likewise.
	(pkl_transl_pr_func): Likewise.
	(pkl_transl_ps_func_arg): Likewise.
	(pkl_transl_ps_decl): Likewise.
	(pkl_transl_pr_struct_type_field): Likewise.
	(pkl_transl_pr_type_struct): Likewise.
	(pkl_transl_ps_type_struct): Likewise.
	* libpoke/pkl-anal.c (pkl_anal1_ps_var): Move checks that require
	lexical depth knowledge to pkl_transl_ps_var.
	* libpoke/pkl.c (rest_of_compilation): Rename passes and add
	transl.
	* libpoke/pkl-gen.c (pkl_gen_ps_var): Emit an ICE if a lexical
	variable in a VAR node is not set.
	* etc/hacking.org (Lexical Handling): New section.
	* HACKING: Regenerate.