pickles: remove PE.opt_hdr.loader_flags == 0 constraint
As discussed on IRC, even though the specification says that this field
"must be zero", .dll files compiled with dotnet(1) on Linux seem to have
this field set to 4096 (0x00001000) instead of 0, and, analysing the
.dll files I have installed on my PC (mostly from wine), although most
of them have this field set to 0, a non-insignificant number of them has
it set to 4096 too.
It is probably a good idea to remove this constraint then.

2024-07-08  Emanuele Torre  <torreemanuele6@gmail.com>

	* pickles/pe.pk (PE_Opt_Hdr): loader_flags is not zero in many
	instances.  Adapt constraint accordingly.