pickles,utils: add pickle and util for BPF CO-RE
BPF Compile Once - Run Everywhere (CO-RE) is a type-based relocation
mechanism which is used in the Linux kernel to patch BPF programs at
load time, in order to accomodate for differences in kernel data
structures between the kernel version on which the BPF program was
compiled and that on which it is being loaded.

CO-RE information is stored in the .BTF.ext section, but relies on
heavily on information from an accompanying BTF section.  Note that
.BTF.ext information is generally BPF-specific, while BTF is not.

bpfcore.pk provides some utilities for working with BPF programs which
have CO-RE relocations.  In particular it provides some routines for
human-readable dumps of the CO-RE relocation information.

pk-bpfcoredump is a utility tool to dump BPF CO-RE relocation
information in a human-readable manner, useful for debugging.

2025-03-03  David Faust  <david.faust@oracle.com>

	* doc/poke.texi (Object Formats): Add entry for BPF.
	(BPF): New node describing the various BPF-related pickles.
	* pickles/bpfcore.pk: New pickle for dealing with BPF CO-RE.
	* pickles/btf.pk (BTF_Type): Minor cleanup and modernization.
	(BTF_Section): Add several new utility methods.
	* pickles/btf-dump.pk (btf_dump_type, btf_dump): Small adaptations
	for above changes.
	* pickles/Makefile.am (dist_pickles_DATA): Add `bpfcore.pk'.
	* utils/Makefile.am (bin_SCRIPTS): Add `pk-bpfcoredump'.
	(EXTRA_DIST): Add `pk-bpfcoredump.in'.
	(pk-bpfcoredump): Add rule for `pk-bpfcoredump.in`.
	* utils/pk-bpfcoredump.in: New Poke script for dumping BPF CO-RE
	relocations.