pvm: add dirty flag and ios ref to pvm_val
Add a new flag, dirty_p, to pvm_mapinfo.  This flag will be checked when
remapping: if a value is not dirty then the underlying bits it is mapped
to are assumed to not have changed since it was last (re-)mapped, so the
remap will be skipped.

Also add an io member to pvm_mapinfo, giving the mapped value a
reference back to the ios in which it is mapped.  It is not used in this
commit, but will be needed to properly deregister the range information
for mapped values from their finalizers in the GC.

2025-06-25  David Faust  <david.faust@oracle.com>

	* libpoke/pvm-val.c (pvm_val_set_dirty): New.
	* libpoke/pvm-val.h (PVM_MAPINFO_DIRTY_P, PVM_MAPINFO_IOS_PTR) New macros.
	(struct pvm_mapinfo): Add dirty_p and io members.
	(PVM_VAL_ARR_DIRTY_P, PVM_VAL_ARR_IOS_PTR, PVM_VAL_SCT_DIRTY)
	(PVM_VAL_SCT_IOS_PTR, PVM_VAL_SET_IOS_PTR, PVM_VAL_DIRTY_P)
	(PVM_VAL_SET_DIRTY_P): New macros.
	* libpoke/pvm.jitter (msetios): Also set ios ptr in the mapped value.
	(mgetd, msetd, mcleard): New instructions.