pvm: add ios range table and interface
This commit introduces the range table to IOS.  Logically, the range
table tracks the intervals associated with every value mapped in the
IOS.  This will be used to identify what currently-mapped values may be
dirtied by a write to an IOS, and need to be remapped.

When a value is mapped in an IOS, it is registered in the range table of
that IOS.  When it is garbage collected, the finalizer deregisters it
from the range table.

The range table is implemented as an augmented interval tree, a form of
binary search tree.  Nodes in the tree are ordered by the bottom end of
the interval where the value is mapped (i.e., its offset), and
additionally store the upper end of the interval and a reference to the
pvm_value mapped there.  As an optimization each node also stores a
"high-water" mark for the highest high-end of any interval in it or its
children, which provides a fast way to prune entire subtrees when
checking interval overlaps.

There is one TODO here: the interval tree does not yet do any form of
self-balancing, so depending on workload it can (and does) decay into a
linked list, for example if values in an IOS are mapped at monotonically
increasing offsets.

2025-06-25  David Faust  <david.faust@oracle.com>

	* libpoke/ios.c (struct interval_node, struct range_table): New.
	(struct ios): Add ranges member.
	(ios_open): Allocate and initialize range table for the ios.
	(ios_close): Empty and free the range table.
	(ios_write_int, ios_write_uint): Mark the affected range dirty.
	(interval_mknode, interval_tree_insert, interval_tree_lookup)
	(recalc_max)
	(interval_tree_min, replace_node, interval_tree_delete)
	(interval_tree_remove, interval_node_overlap, interval_tree_mark)
	(interval_tree_mark_all, interval_tree_empty): New static
	functions implementing the interval tree.
	(range_table_remove, range_table_empty, range_table_insert)
	(range_table_mark_dirty, range_table_mark_dirty_all): New
	functions implenting the interface to the range table.
	(ios_register_range, ios_get_ranges, ios_deregister_range)
	(ios_mark_dirty_range, ios_mark_dirty_all): New functions
	implementing the ios range interface.
	* libpoke/ios.h (ios_register_range, ios_deregister_range)
	(ios_mark_dirty_range, ios_get_ranges, ios_mark_dirty_all):
	Declare them here.
	* libpoke/pvm-alloc.c (pvm_alloc_finalize_struct)
	(pvm_alloc_finalize_array): New.
	(pvm_alloc_arr, pvm_alloc_sct): New.
	* libpoke/pvm-alloc.h (pvm_alloc_arr, pvm_alloc_sct): Declarations here.
	* libpoke/pvm-val.c (pvm_make_array): Use pvm_alloc_arr.
	(pvm_make_sct): Likewise use pvm_alloc_sct.
	* libpoke/pvm.c (pvm_shutdown): Trigger a GC pass between
	finalizing values and calling ios_shutdown.
	* libpoke/pvm.jitter (ioregval, iodirtyall): New.