pvm,pkl: mark all mapped values dirty if endianness changes
Besides writes to the IOS, there is one additional case where we must
mark values dirty, namely when the VM endianness changes.  In that case,
all values must be marked dirty, so that they will be remapped in the
new endianness.  Some care is required for structs, since struct field
constraints can cause the endianness to change part way through mapping
a struct.

2025-06-25  David Faust  <david.faust@oracle.com>

	* libpoke/pkl-gen.pks (RAS_FUNCTION_STRUCT_MAPPER): Mark newly
	mapped struct dirty if endianness changed while mapping its
	fields.
	* libpoke/pkl-rt.pk (dirty, dirty_ios): New.
	* libpoke/pvm.c (pvm_set_endian): Mark every mapped value in every
	ios dirty when the VM endianness changes.
	* libpoke/pvm.jitter (popend): Likewise.