pkl: set trimmed array offset before adding elements
The array trimmer did not previously set the offset of the new trimmed
array before adding elements to it.  This was ok, since the new array
would be remapped, during which the element offsets would be
recalculated.  However, this remap is unnecessary from a data-freshness
perspective, and won't be done if the new array is not considered dirty.
Set the offset of the trimmed array before adding elements, so that the
elements inherit the correct base offset when computing each element
offset.

2025-06-25  David Faust  <david.faust@oracle.com>

	* libpoke/pkl-gen.pks (ATRIM): Set offset of new array before
	adding elements so the elements inherit the proper base offset.