pkl: skip remap if value is not dirty
This commit changes remap and aremap to check the dirty_p flag of the
value before performing the actual remap.  If the value is not dirty,
then the remap is skipped.

The remap macro now pushes an additional value to the stack indicating
whether (1) or not (0) the remap was actually necessary and thus
performed.

This new value is of particular interest to the passes in pkl-gen.c
which can issue remaps.  If/when a remap happens, the returned value is
actually a new distinct value in the poke virtual machine, and
references to the old value may remain, which can significantly delay
garbage collection of the old value.  Since these values are now also
tracked in the IOS range tables, repeated remaps cause the range tables
to grow excessively.

To alleviate this, gen_ps_var, gen_pr_indexer and
gen_ps_struct_ref (which may issue the remaps) now check the remapped_p
flag pushed on the stack by the remap operation, and replace the old
value with the new one in the corresponding frame if the remap happened.
This reduces dangling references to the old values so they can be
garbage collected and deregistered from the IOS range table much sooner.

Co-authored-by: Jose E. Marchesi  <jemarch@gnu.org>

2025-06-25  David Faust  <david.faust@oracle.com>
            Jose E. Marchesi  <jemarch@gnu.org>

	* libpoke/pkl-asm.pks (RAS_MACRO_AREMAP): Push a new REMAPPED_P on
	the stack along with the NVAL to indicate whether or not the value
	was actually remapped.
	(RAS_MACRO_REMAP): Likewise. Skip remapping a value if its dirty
	flag is not set.
	* libpoke/pkl-gen.c (pkl_gen_ps_var): Replace old value in
	corresponding frame if it was remapped.
	(pkl_gen_pr_indexer): Likewise.
	(pkl_gen_ps_struct_ref): Likewise.
	(pkl_gen_ps_op_remap): Accomodate new value on stack after remap.
	* libpoke/pkl-gen.pks (RAS_MACRO_AOERF): Likewise. Correct
	macro-level comment on stack state at finish.